added Yaml & Kaitai Struct decoder, started adding Kaitai Template Controller
Kaitai uses Yaml files to describe binary structures. I've created a Yaml decoder based on libYaml (only works if it's installed, doesn't do no hard, if its missing).
The decoded Yaml is then converted in to HFKsyObject instances (HFKsyUserDefinedType being the root object). It basically converts the dictionary data to objects.
The HFKaitaiTemplateController is another HFTemplateController, sibling of HFTclTemplateController. It takes the .ksy template files, parses them and applies them to the current data-file.
That last part is not fully working yet, but the basics are there and work nicely already.